requirements.txt
aiogram==3.6.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
APScheduler==3.10.4
